Fibonacci Series by recursive function and few error messages with red color are added to my program.
Assalamu alaikum sir.

This program is for experiment and learning. Sometimes different features are added/removed to observe the output. It may seem some features are unnecessary or some lines may be repeated, but it should be kept in mind that it is a method of learning.

Thank you sir.