Tower of Hanoi solved with recursive function and error handling is added two or tree places.
If input value is invalid, an error message is helpful for the user. And it also shows the skill of the programmer. When I felt the requirement of error handling, then I added those immediate after. That is why somewhere in my early portions of code, there still may remain some inconvenience. However, I am learning and that is positive symbol for programming.